requirements/base.txt
-i https://pypi.org/simple
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.2.1 # pyup: != 11.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.9.0  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.1  # https://github.com/redis/hiredis-py
celery==5.5.2  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat # https://github.com/celery/django-celery-beat
# django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==5.1.9 # pyup: < 5.2 # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.8.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-tailwind==1.0.3  # https://github.com/django-crispy-forms/crispy-tailwind
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.7.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.28.0  # https://github.com/tfranzel/drf-spectacular
django-webpack-loader==3.1.1  # https://github.com/django-webpack/django-webpack-loader
django-celery-results
# django-celery-results==2.5.0
django-filter==25.1
django-simple-captcha==0.6.1
django-htmx==1.21.0
django-widget-tweaks==1.5.0
django-tables2==2.7.5
django-render-block==0.10
django-admin-sortable2==2.2.4
django-formset-js-improved==0.5.0.3
django-formtools==2.5.1
django-import-export==4.3.5
django-unfold==0.49.1
scrapy-redis==0.9.1
celery-progress==0.5
dateparser==1.2.1
olefile==0.47
django-modeltranslation==0.19.12
django-dynamic-formsets==0.0.8
# bs4==0.0.2
django-ckeditor-5==0.2.17
pyjwt
django-template-partials==24.4

Scrapy==2.13.0
# scrapy-headless-selenium==0.1.2
# selenium==4.28.1

# typing==3.7.4.3
# typing_extensions==4.12.2
# webdriver-manager
bz2file==0.98
django-render-block==0.10
# scrapy-impersonate==1.4.1
django-autocomplete-light
beautifulsoup4
html5lib
lxml
twisted-iocpsupport
wheel
# django-dynamic-datatb
# webdriver-manager
extruct
pydantic
typing
# selenium-wire
blinker==1.7.0
# selenium-stealth
# selenium-driverless
# asynciolimiter
scrapy-playwright
djangorestframework-simplejwt[crypto]
nodriver
dj-rest-auth[with_social]
# scrapy-nodriver
supabase


